services/api/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
pydantic>=2.7
pydantic-settings>=2.1.0
pymysql==1.1.0
python-dotenv==1.0.0
cryptography==41.0.7
sqlalchemy>=2.0.0
python-logging-loki>=0.3.1
holidays>=0.40
fastapi-cache2>=0.2.1
redis>=5.0.0
bcrypt>=4.0.0
PyJWT>=2.8.0
google-auth>=2.25.0
requests>=2.31.0
